Groundstation/relay/ext/pi-protocol/python/requirements.txt
Jinja2==3.1.4
PyYAML==6.0.1
semver==3.0.1
